linked journal entries with users

+++ src/main/java/com/example/journalApp/services/JournalService.java
@@ -1,6 +1,7 @@
 package com.example.journalApp.services;
 
 import com.example.journalApp.entity.JournalEntry;
+import com.example.journalApp.entity.User;
 import com.example.journalApp.repository.JournalRepository;
 import org.bson.types.ObjectId;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -15,20 +16,38 @@ public class JournalService {
     @Autowired
     private JournalRepository journalRepository;
 
+    @Autowired
+    private UserService userService;
+
+    public void saveEntry (JournalEntry entry, String username) {
+        User user = userService.getUserByUsername(username);
+        if(user != null) {
+            entry.setDate(LocalDateTime.now());
+            JournalEntry savedEntry = journalRepository.save(entry);
+            user.getJournalEntries().add(savedEntry);
+            userService.saveUser(user);
+        }
+    }
+
     public void saveEntry (JournalEntry entry) {
-        entry.setDate(LocalDateTime.now());
         journalRepository.save(entry);
     }
 
-    public List<JournalEntry> getAll() {
-        return journalRepository.findAll();
+    public List<JournalEntry> getAll(String username) {
+        User user = userService.getUserByUsername(username);
+        return user.getJournalEntries();
     }
 
     public Optional<JournalEntry> findById (ObjectId id) {
         return journalRepository.findById(id);
     }
 
-    public void deleteById (ObjectId id) {
-        journalRepository.deleteById(id);
+    public void deleteById (ObjectId id, String username) {
+        User user = userService.getUserByUsername(username);
+        if (user != null) {
+            user.getJournalEntries().removeIf(x -> x.getId().equals(id));
+            userService.saveUser(user);
+            journalRepository.deleteById(id);
+        }
     }
 }
